Move EnvironmentsHelper into common project
