+++ src/Jammehcow.YosherBot.Common/Properties/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Jammehcow.YosherBot.UnitTest")]
