+++ src/Jammehcow.YosherBot.Console/BotStartup.Init.cs
@@ -3,6 +3,7 @@ using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using Jammehcow.YosherBot.Command.Administration;
+using Jammehcow.YosherBot.Common.Helpers.Environment;
 using Jammehcow.YosherBot.Console.Extensions;
 using Jammehcow.YosherBot.Console.Helpers;
 using Microsoft.Extensions.Logging;
