+++ src/Jammehcow.YosherBot.Console/Program.cs
@@ -4,6 +4,7 @@ using Autofac;
 using Autofac.Extensions.DependencyInjection;
 using Discord;
 using Discord.WebSocket;
+using Jammehcow.YosherBot.Common.Helpers.Environment;
 using Jammehcow.YosherBot.Console.Extensions;
 using Jammehcow.YosherBot.Console.Helpers;
 using Microsoft.Extensions.DependencyInjection;
