+++ src/Jammehcow.YosherBot.UnitTest/EnvironmentsHelperTests.cs
@@ -1,4 +1,5 @@
-﻿using Jammehcow.YosherBot.Console.Helpers;
+﻿using Jammehcow.YosherBot.Common.Helpers.Environment;
+using Jammehcow.YosherBot.Console.Helpers;
 using NUnit.Framework;
 
 namespace Jammehcow.YosherBot.UnitTest
