+++ src/Jammehcow.YosherBot.UnitTest/TestHelper.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Jammehcow.YosherBot.Common.Helpers.Environment;
 using Jammehcow.YosherBot.Console.Helpers;
 
 namespace Jammehcow.YosherBot.UnitTest
